refactor: split ItemSkills class into smaller classes
